if you're gonna pull don't pot
anguish. sorrow. misery. these are but a few qualities that enrapture our team in our final moments. was there ever a moment where we would truly be successful? nay, i don't believe there was. but i believe that we all understood this, and in fact, this might be the saving grace of our project, oh you that was doomed to fail.

for as sisyphus moves his boulder, so too will we try and update the scale of a particular set of keybinds. so too, will we design a game that is 90% walking around and changing the background track. so too, we will design a game that is 100% carried by its artistic integrity. for its true bohemenian nature reflects that of our state as college students.

truly this reminds me of Dante Aligheri meeting with the famed poet Virgil in Limbo outside the caves of Italy, forced into exile and mourning his one true love. truly, have we suffered the same as those selfsame poets? truly, are we the most tortured poets?

and, damn, if we don't live in a society.

i'm gonna go get my monitor.